feat: add Billing and Settings pages with navigation and subscription details
